Fix CLI UX.

+++ src/Command/DeleteCommand.php
@@ -2,14 +2,12 @@
 
 namespace Drupal\static_generator\Command;
 
+use Drupal\Console\Annotations\DrupalCommand;
+use Drupal\Console\Core\Command\ContainerAwareCommand;
 use Drupal\static_generator\StaticGenerator;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
-use Drupal\Console\Core\Command\ContainerAwareCommand;
-use Drupal\Console\Annotations\DrupalCommand;
-
 
 /**
  * Class DeleteAllCommand.
@@ -66,29 +64,44 @@ class DeleteCommand extends ContainerAwareCommand {
    * @throws \Exception
    */
   protected function execute(InputInterface $input, OutputInterface $output) {
-    if ($input->getOption('pages')) {
-      $answer = $this->getIo()->ask('Delete all pages (yes/no)? ');
-      if (strtolower($answer) == 'yes') {
-        $elapsed_time = $this->staticGenerator->deletePages();
-        $this->getIo()
-          ->info('Delete all generated pages completed, elapsed time: ' . $elapsed_time . ' seconds.');
+    // Get confirmation option.
+    $confirmed = $input->getOption('yes');
+
+    // Initiatlize answer var in case confirmation is not set.
+    $answer = 'no';
+
+    // If not confirmed with '--yes', ask.
+    if (!$confirmed) {
+      if ($input->getOption('pages')) {
+        $answer = $this->getIo()->ask('Delete all pages (yes/no)? ');
+        if (strtolower($answer) == 'yes') {
+          $elapsed_time = $this->staticGenerator->deletePages();
+          $this->getIo()
+            ->info('Delete all generated pages completed, elapsed time: ' . $elapsed_time . ' seconds.');
+        }
+      } elseif ($input->getOption('esi')) {
+        $answer = $this->getIo()->ask('Delete all ESIs (yes/no)?');
+        if (strtolower($answer) == 'yes') {
+          $elapsed_time = $this->staticGenerator->deleteEsi();
+          $this->getIo()
+            ->info('Delete all generated ESIs completed, elapsed time: ' . $elapsed_time . ' seconds.');
+        }
+      } else {
+        $answer = $this->getIo()->ask('Delete all ESIs, files and pages (yes/no)?');
+        if (strtolower($answer) == 'yes') {
+          $elapsed_time = $this->staticGenerator->deleteAll();
+          $this->getIo()
+            ->info('Delete all generated ESIs, files and pages completed, elapsed time: ' . $elapsed_time . ' seconds.');
+        }
       }
     }
-    elseif ($input->getOption('esi')) {
-      $answer = $this->getIo()->ask('Delete all ESIs (yes/no)?');
-      if (strtolower($answer) == 'yes') {
-        $elapsed_time = $this->staticGenerator->deleteEsi();
-        $this->getIo()
-          ->info('Delete all generated ESIs completed, elapsed time: ' . $elapsed_time . ' seconds.');
-      }
+
+    // If confirmed, delete all.
+    if ($confirmed || strtolower($answer) == 'yes') {
+      $elapsed_time = $this->staticGenerator->deleteAll();
+      $this->getIo()
+        ->info('Delete all generated ESIs, files and pages completed, elapsed time: ' . $elapsed_time . ' seconds.');
     }
-    else {
-      $answer = $this->getIo()->ask('Delete all ESIs, files and pages (yes/no)?');
-      if (strtolower($answer) == 'yes') {
-        $elapsed_time = $this->staticGenerator->deleteAll();
-        $this->getIo()
-          ->info('Delete all generated ESIs, files and pages completed, elapsed time: ' . $elapsed_time . ' seconds.');
-      }
-    }
+
   }
 }
